life_sim_bot_complete/requirements.txt

discord.py
motor
python-dotenv
